+++ jest.esm.mjs
@@ -28,8 +28,7 @@ export default {
   transform: {},
   restoreMocks: true,
   moduleNameMapper: {
-    '^@okta/okta-auth-js$': OktaAuth,
-    '^broadcast-channel$': '<rootDir>/node_modules/broadcast-channel/dist/es5node/index.js'
+    '^@okta/okta-auth-js$': OktaAuth
   },
   extensionsToTreatAsEsm: ['.ts'],
   testPathIgnorePatterns: [],